+++ src/common/components/Checkbox/Checkbox.module.css
@@ -11,14 +11,15 @@
   width: 16px;
   height: 16px;
 }
-.checkboxContainer {
+.checkboxLabel {
   display: inline-block;
   width: 100%;
   min-height: 16px;
   position: relative;
+  padding: 0 0 16px 24px;
   cursor: pointer;
 }
-.checkboxContainer::before {
+.checkboxLabel::before {
   content: "";
   display: inline-block;
   position: absolute;
@@ -31,7 +32,7 @@
   border-radius: 2px;
   background-color: #fff;
 }
-.checkboxContainer::after {
+.checkboxLabel::after {
   content: "";
   display: inline-block;
   position: absolute;
@@ -54,10 +55,10 @@
   display: none;
   position: absolute;
 }
-.checkboxInput:checked + .checkboxContainer::before {
+.checkboxInput:checked + .checkboxLabel::before {
   background-color: #459df5;
 }
-.checkboxInput:checked + .checkboxContainer::after {
+.checkboxInput:checked + .checkboxLabel::after {
   opacity: 1;
   border: none;
 }
