+++ src/common/components/Dropdown/Dropdown.module copy.css
@@ -1,0 +1,259 @@
+.button {
+  background-repeat: no-repeat;
+  background-color: unset;
+  border: unset;
+  cursor: pointer;
+  border-radius: 4px;
+  display: flex;
+  flex-wrap: nowrap;
+}
+.button:focus {
+  outline: none;
+}
+.input {
+  width: 100%;
+  min-height: 32px;
+  border-radius: 4px;
+}
+.input:focus {
+  outline: none;
+}
+.input::-moz-placeholder {
+  color: #bad8f5;
+}
+.input::-webkit-input-placeholder {
+  color: #bad8f5;
+}
+.buttonThemeBlue {
+  border: 1px solid #459df5;
+  background-color: #459df5;
+  color: #fff;
+}
+.buttonThemeBlue .icon {
+  fill: #fff;
+}
+.buttonThemeLink {
+  color: #459df5;
+}
+.buttonThemeLink .icon {
+  fill: #459df5;
+}
+.buttonThemeText .icon {
+  fill: #000;
+}
+.buttonSizeDefault {
+  padding: 8px 16px;
+}
+.buttonSizeSmall {
+  padding: 4px 8px;
+}
+.buttonSizeDefault + .buttonIconOnly {
+  padding: 8px 8px;
+}
+.buttonSizeSmall + .buttonIconOnly {
+  padding: 4px 4px;
+}
+.icon {
+  width: 16px;
+  height: 16px;
+}
+.icon + .buttonText {
+  margin-left: 8px;
+}
+.dropdownWrapper {
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+  margin-bottom: 96px;
+}
+.dropdownWrapper:last-child {
+  margin-bottom: 0;
+}
+.dropdownBlockCheckbox {
+  margin-right: 48px;
+}
+.dropdownBlockCheckbox:last-child {
+  margin-right: 0;
+}
+.formBlockSelectPage {
+  margin-right: 48px;
+}
+.formBlockSelectPage,
+.formBlockDeleteRecord {
+  margin-right: 48px;
+}
+.checkboxDropdown,
+.radioDropdown {
+  width: 160px;
+  height: 208px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: flex-start;
+  padding: 16px;
+  border-radius: 8px;
+  background: #fff;
+}
+input[type="radio"]:checked + label {
+  color: #459df5;
+}
+input[type="checkbox"]:checked + label {
+  color: #459df5;
+}
+.selectPage {
+  width: 160px;
+  height: 88px;
+  padding: 16px;
+  border-radius: 8px;
+  background: #fff;
+}
+.deleteRecord,
+.selectTheme {
+  width: 160px;
+  height: 112px;
+  padding: 16px;
+  border-radius: 8px;
+  background: #fff;
+}
+.selectPageInput {
+  border: 1px solid #459df5;
+  padding: 0 4px 0 8px;
+}
+.selectPageTitle,
+.deleteRecordTitle,
+.selectThemeTitle {
+  display: inline-block;
+  margin-bottom: 8px;
+}
+.deleteRecordButton,
+.selectThemeButton {
+  margin-bottom: 8px;
+}
+.deleteTecordButton:last-child,
+.selectThemeButton:last-child {
+  margin-bottom: 0;
+}
+.buttonFullWidth {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+}
+.selectThemeIconUnpushed {
+  fill: #459df5;
+}
+.selectThemeIconPushed {
+  fill: #fff;
+}
+.checkboxContent,
+.radioContent {
+  width: 16px;
+  height: 16px;
+}
+.checkboxContainer {
+  display: inline-block;
+  width: 100%;
+  position: relative;
+  padding: 0 0 16px 24px;
+  cursor: pointer;
+}
+.checkboxContainer::before {
+  content: "";
+  display: inline-block;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 1;
+  height: 16px;
+  width: 16px;
+  border: 1px solid #bad8f5;
+  border-radius: 2px;
+  background-color: #fff;
+}
+.checkboxContainer::after {
+  content: "";
+  display: inline-block;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 2;
+  height: 16px;
+  width: 16px;
+  border: 1px solid #bad8f5;
+  border-radius: 2px;
+  background-image: url(../../../icons/checkmark.svg);
+  background-repeat: no-repeat;
+  opacity: 0;
+}
+.radioContainer {
+  display: inline-block;
+  width: 100%;
+  position: relative;
+  padding: 0 0 16px 24px;
+  cursor: pointer;
+}
+.radioContainer::before {
+  content: "";
+  display: inline-block;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 1;
+  height: 16px;
+  width: 16px;
+  border: 1px solid #bad8f5;
+  border-radius: 50%;
+  background-color: #fff;
+}
+.radioContainer::after {
+  content: "";
+  display: inline-block;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 2;
+  height: 16px;
+  width: 16px;
+  border-radius: 50%;
+  background-image: url(../../../icons/dot.svg);
+  background-repeat: no-repeat;
+  opacity: 0;
+}
+.checkbox,
+.radio {
+  height: 16px;
+  min-width: 16px;
+}
+.checkboxInput,
+.radioInput {
+  display: none;
+  position: absolute;
+}
+.checkboxInput:checked + .checkboxContainer::before {
+  background-color: #459df5;
+}
+.checkboxInput:checked + .checkboxContainer::after {
+  opacity: 1;
+  border: none;
+}
+.radioInput:checked + .radioContainer::before {
+  background-color: #459df5;
+}
+.radioInput:checked + .radioContainer::after {
+  opacity: 1;
+  border: none;
+}
+.checkboxDropdownItem,
+.radioDropdownItem,
+.checkboxDropdownCheckbox,
+.radioDropdownRadio {
+  width: 100%;
+}
+.radioContainerNoIco {
+  padding-left: 0;
+}
+.radioContainerNoIco::before {
+  display: none;
+}
+.radioContainerNoIco::after {
+  display: none;
+}
