+++ src/common/components/Dropdown/Dropdown.module.css
@@ -1,89 +1,50 @@
-.button {
-  background-repeat: no-repeat;
-  background-color: unset;
-  border: unset;
-  cursor: pointer;
-  border-radius: 4px;
+.dropdownBlock {
+  margin-right: 48px;
+}
+.dropdownBlock:last-child {
+  margin-right: 0;
+}
+.Dropdown {
+  width: 160px;
+  height: 208px;
   display: flex;
-  flex-wrap: nowrap;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: flex-start;
+  padding: 16px;
+  border-radius: 8px;
+  background: #fff;
 }
-.button:focus {
-  outline: none;
+
+.ui-kit__block_dropdown {
+  width: 674px;
+  min-width: 674px;
+  padding: 48px;
 }
-.input {
-  width: 100%;
-  min-height: 32px;
-  border-radius: 4px;
-}
-.input:focus {
-  outline: none;
-}
-.input::-moz-placeholder {
-  color: #bad8f5;
-}
-.input::-webkit-input-placeholder {
-  color: #bad8f5;
-}
-.buttonThemeBlue {
-  border: 1px solid #459df5;
-  background-color: #459df5;
-  color: #fff;
-}
-.buttonThemeBlue .icon {
-  fill: #fff;
-}
-.buttonThemeLink {
-  color: #459df5;
-}
-.buttonThemeLink .icon {
-  fill: #459df5;
-}
-.buttonThemeText .icon {
-  fill: #000;
-}
-.buttonSizeDefault {
-  padding: 8px 16px;
-}
-.buttonSizeSmall {
-  padding: 4px 8px;
-}
-.buttonSizeDefault + .buttonIconOnly {
-  padding: 8px 8px;
-}
-.buttonSizeSmall + .buttonIconOnly {
-  padding: 4px 4px;
-}
-.icon {
-  width: 16px;
-  height: 16px;
-}
-.icon + .buttonText {
-  margin-left: 8px;
-}
-.dropdownWrapper {
+.dropdown__wrapper {
   display: flex;
   justify-content: flex-start;
   align-items: flex-start;
   margin-bottom: 96px;
 }
-.dropdownWrapper:last-child {
+.dropdown__wrapper:last-child {
   margin-bottom: 0;
 }
-.dropdownBlockCheckbox {
+.dropdown-block__checkbox {
   margin-right: 48px;
 }
-.dropdownBlockCheckbox:last-child {
+.dropdown-block__checkbox:last-child {
   margin-right: 0;
 }
-.formBlockSelectPage {
+.form-block__select-page {
   margin-right: 48px;
 }
-.formBlockSelectPage,
-.formBlockDeleteRecord {
+.form-block__select-page,
+.form-block__delete-record {
   margin-right: 48px;
 }
-.checkboxDropdown,
-.radioDropdown {
+.checkbox-dropdown,
+.radio-dropdown {
   width: 160px;
   height: 208px;
   display: flex;
@@ -100,160 +61,56 @@ input[type="radio"]:checked + label {
 input[type="checkbox"]:checked + label {
   color: #459df5;
 }
-.selectPage {
+.select-page {
   width: 160px;
   height: 88px;
   padding: 16px;
   border-radius: 8px;
   background: #fff;
 }
-.deleteRecord,
-.selectTheme {
+.delete-record,
+.select-theme {
   width: 160px;
   height: 112px;
   padding: 16px;
   border-radius: 8px;
   background: #fff;
 }
-.selectPageInput {
+.select-page__input {
   border: 1px solid #459df5;
   padding: 0 4px 0 8px;
 }
-.selectPageTitle,
-.deleteRecordTitle,
-.selectThemeTitle {
+.select-page__title,
+.delete-record__title,
+.select-theme__title {
   display: inline-block;
   margin-bottom: 8px;
 }
-.deleteRecordButton,
-.selectThemeButton {
+.delete-record__button,
+.select-theme__button {
   margin-bottom: 8px;
 }
-.deleteTecordButton:last-child,
-.selectThemeButton:last-child {
+.delete-record__button:last-child,
+.select-theme__button:last-child {
   margin-bottom: 0;
 }
-.buttonFullWidth {
+.button_full-width {
   width: 100%;
   display: flex;
   justify-content: center;
 }
-.selectThemeIconUnpushed {
+.select-theme__icon_unpushed {
   fill: #459df5;
 }
-.selectThemeIconPushed {
+.select-theme__icon_pushed {
   fill: #fff;
 }
-.checkboxContent,
-.radioContent {
-  width: 16px;
-  height: 16px;
-}
-.checkboxContainer {
-  display: inline-block;
-  width: 100%;
-  position: relative;
-  padding: 0 0 16px 24px;
-  cursor: pointer;
-}
-.checkboxContainer::before {
-  content: "";
-  display: inline-block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 1;
-  height: 16px;
-  width: 16px;
-  border: 1px solid #bad8f5;
-  border-radius: 2px;
-  background-color: #fff;
-}
-.checkboxContainer::after {
-  content: "";
-  display: inline-block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 2;
-  height: 16px;
-  width: 16px;
-  border: 1px solid #bad8f5;
-  border-radius: 2px;
-  background-image: url(../../../icons/checkmark.svg);
-  background-repeat: no-repeat;
-  opacity: 0;
-}
-.radioContainer {
-  display: inline-block;
-  width: 100%;
-  position: relative;
-  padding: 0 0 16px 24px;
-  cursor: pointer;
-}
-.radioContainer::before {
-  content: "";
-  display: inline-block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 1;
-  height: 16px;
-  width: 16px;
-  border: 1px solid #bad8f5;
-  border-radius: 50%;
-  background-color: #fff;
-}
-.radioContainer::after {
-  content: "";
-  display: inline-block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 2;
-  height: 16px;
-  width: 16px;
-  border-radius: 50%;
-  background-image: url(../../../icons/dot.svg);
-  background-repeat: no-repeat;
-  opacity: 0;
-}
-.checkbox,
-.radio {
-  height: 16px;
-  min-width: 16px;
-}
-.checkboxInput,
-.radioInput {
-  display: none;
-  position: absolute;
-}
-.checkboxInput:checked + .checkboxContainer::before {
-  background-color: #459df5;
-}
-.checkboxInput:checked + .checkboxContainer::after {
-  opacity: 1;
-  border: none;
-}
-.radioInput:checked + .radioContainer::before {
-  background-color: #459df5;
-}
-.radioInput:checked + .radioContainer::after {
-  opacity: 1;
-  border: none;
-}
-.checkboxDropdownItem,
-.radioDropdownItem,
-.checkboxDropdownCheckbox,
-.radioDropdownRadio {
-  width: 100%;
-}
-.radioContainerNoIco {
+.radio__container_no-ico {
   padding-left: 0;
 }
-.radioContainerNoIco::before {
+.radio__container_no-ico::before {
   display: none;
 }
-.radioContainerNoIco::after {
+.radio__container_no-ico::after {
   display: none;
 }
