+++ src/common/components/Radio/Radio.jsx
@@ -3,52 +3,36 @@ import cx from "classnames";
 import styles from "./Radio.module.css";
 
 export const Radio = ({
-  children,
-  className,
-  value,
+  id,
+  name,
+  type,
   onChange,
-  isRadioChecked,
+  labelValue,
+  isRadioChecked = false,
+  className,
   ...props
 }) => {
-  const [radioOneChecked, setRadioOneChecked] = useState("");
-  const [radioTwoChecked, setRadioTwoChecked] = useState("checked");
+  const [radioChecked, setRadioChecked] = useState(isRadioChecked);
 
-  const handleRadioOne = ({ target: { checked } }) => {
-    setRadioOneChecked(checked);
-  };
-
-  const handleRadioTwo = ({ target: { checked } }) => {
-    setRadioTwoChecked(checked);
+  const handleRadio = ({ target: { checked } }) => {
+    setRadioChecked(checked);
   };
 
   return (
-    <>
-      <div className={cx(styles.radio, styles.blockRadio)}>
-        <div className={styles.radioContent}>
-          <input
-            className={styles.radioInput}
-            name="radio-name"
-            type="radio"
-            checked={radioOneChecked}
-            onChange={handleRadioOne}
-            id="radio_1"
-          />
-          <label className={styles.radioContainer} htmlFor="radio_1"></label>
-        </div>
+    <div className={cx(styles.radio, styles.blockRadio)}>
+      <div className={styles.radioContent}>
+        <input
+          id={id}
+          name={name}
+          type={type}
+          checked={radioChecked}
+          onChange={handleRadio}
+          className={styles.radioInput}
+        />
+        <label className={styles.radioLabel} htmlFor={id}>
+          {labelValue}
+        </label>
       </div>
-      <div className={cx(styles.radio, styles.blockRadio)}>
-        <div className={styles.radioContent}>
-          <input
-            className={styles.radioInput}
-            name="radio-name"
-            type="radio"
-            checked={radioTwoChecked}
-            onChange={handleRadioTwo}
-            id="radio_2"
-          />
-          <label className={styles.radioContainer} htmlFor="radio_2"></label>
-        </div>
-      </div>
-    </>
+    </div>
   );
 };
