+++ src/icons/index.js
@@ -1,0 +1,6 @@
+export { ReactComponent as Checkmark } from "./checkmark.svg";
+export { ReactComponent as Dot } from "./dot.svg";
+export { ReactComponent as DotLarge } from "./dot-large.svg";
+export { ReactComponent as Locked } from "./locked.svg";
+export { ReactComponent as Search } from "./search.svg";
+export { ReactComponent as XMedium } from "./x-medium.svg";
